+++ include/pfnet/pfnet.h
@@ -42,7 +42,6 @@
 #include "constr_LINPF.h"
 #include "constr_FIX.h"
 #include "constr_LBOUND.h"
-#include "constr_NBOUND.h"
 #include "constr_PAR_GEN_P.h"
 #include "constr_PVPQ_SWITCHING.h"
 #include "constr_GEN_RAMP.h"
@@ -57,6 +56,6 @@
 #include "constr_CFUNC.h"
 
 // Heuristics
-#include "heur_PVPQ.h"
+#include "heur_PVPQ_SWITCHING.h"
 
 #endif